+++ src/autowiring/CMakeLists.txt
@@ -25,7 +25,6 @@ set(Autowiring_SRCS
   auto_out.h
   auto_prev.h
   auto_signal.h
-  auto_signal.cpp
   auto_tuple.h
   AutoCurrentPacketPusher.h
   AutoFilterDescriptor.h
@@ -134,6 +133,8 @@ set(Autowiring_SRCS
   Parallel.cpp
   registration.h
   SatCounter.h
+  signal.h
+  signal.cpp
   signal_base.h
   SlotInformation.cpp
   SlotInformation.h
