+++ src/autowiring/signal.cpp
